feature(component/login):improved styles and functionalities
