+++ E-commerce/client/pages/login.js
@@ -1,13 +1,9 @@
 import Login from "../components/Login/Login";
-import { SlideFade } from "@chakra-ui/react";
-import { useRouter } from "next/router";
 
 function LoginPage() {
-  const router = useRouter();
+
   return (
-    <SlideFade key={router.route} in="true">
       <Login />
-    </SlideFade>
   );
 }
